+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Features
 - Enable native tests on CI (#204)
 - Migrate to Ethers v6:
-  - Deployment scripts (#187)
+  - Deployment scripts (#187, #203)
   - Tests
     - IexecEscrow (#199)
     - ENSIntegration, IexecOrderManagement, IexecRelay (#195, #199)